# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "シーシャSNSアプリケーションのバックエンドAPI\nこのAPIはシーシャの投稿、ユーザー管理を行います",
        "title": "Go-Shisha API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/illionillion/go-shisha"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "全てのユーザーの一覧を取得します（総数付き）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー一覧取得",
                "responses": {
                    "200": {
                        "description": "ユーザー一覧と総数",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.UsersResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "指定されたIDのユーザー情報を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー詳細取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ユーザー情報",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}/posts": {
            "get": {
                "description": "指定されたユーザーの全ての投稿を取得します（総数付き）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザーの投稿一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "投稿一覧と総数",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.PostsResponse"
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "go-shisha-backend_internal_models.Flavor": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "go-shisha-backend_internal_models.Post": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "slides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shisha-backend_internal_models.Slide"
                    }
                },
                "user": {
                    "$ref": "#/definitions/go-shisha-backend_internal_models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "go-shisha-backend_internal_models.PostsResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "go-shisha-backend_internal_models.Slide": {
            "type": "object",
            "required": [
                "image_url"
            ],
            "properties": {
                "flavor": {
                    "$ref": "#/definitions/go-shisha-backend_internal_models.Flavor"
                },
                "image_url": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "go-shisha-backend_internal_models.User": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "external_url": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "go-shisha-backend_internal_models.UsersResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shisha-backend_internal_models.User"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}